+++ src/core/firebase.app.module.ts
@@ -1,5 +1,5 @@
-import { InjectionToken, NgModule, Optional, VERSION as NG_VERSION, Version } from '@angular/core';
-import { auth, database, firestore, functions, messaging, storage } from 'firebase/app';
+import { InjectionToken, NgModule, Optional, NgZone, VERSION as NG_VERSION, Version } from '@angular/core';
+import { auth, database, messaging, storage, firestore, functions } from 'firebase/app';
 // @ts-ignore (https://github.com/firebase/firebase-js-sdk/pull/1206)
 import firebase from 'firebase/app'; // once fixed can pull in as "default as firebase" above
 
@@ -7,35 +7,47 @@ import firebase from 'firebase/app'; // once fixed can pull in as "default as fi
 export type FirebaseOptions = {[key:string]: any};
 export type FirebaseAppConfig = {[key:string]: any};
 
+// SEMVER drop FirebaseOptionsToken and FirebaseNameOrConfigToken in favor of FIREBASE_OPTIONS and FIREBASE_APP_NAME in next major
 export const FirebaseOptionsToken = new InjectionToken<FirebaseOptions>('angularfire2.app.options');
-export const FirebaseNameOrConfigToken = new InjectionToken<string|FirebaseAppConfig|undefined>('angularfire2.app.nameOrConfig')
+export const FirebaseNameOrConfigToken = new InjectionToken<string|FirebaseAppConfig|undefined>('angularfire2.app.nameOrConfig');
+
+export const FIREBASE_OPTIONS = FirebaseOptionsToken;
+export const FIREBASE_APP_NAME = FirebaseNameOrConfigToken;
 
 export type FirebaseDatabase = database.Database;
 export type FirebaseAuth = auth.Auth;
+// SEMVER analytics.Analytics;
+export type FirebaseAnalytics = any;
 export type FirebaseMessaging = messaging.Messaging;
+// SEMVER performance.Performance
+export type FirebasePerformance = any;
 export type FirebaseStorage = storage.Storage;
 export type FirebaseFirestore = firestore.Firestore;
 export type FirebaseFunctions = functions.Functions;
+// SEMVER remoteConfig.RemoteConfig;
+export type FirebaseRemoteConfig = any;
 
 // Have to implement as we need to return a class from the provider, we should consider exporting
 // this in the firebase/app types as this is our highest risk of breaks
 export class FirebaseApp {
     name: string;
     options: {};
+    analytics: () => FirebaseAnalytics;
     auth: () => FirebaseAuth;
     database: (databaseURL?: string) => FirebaseDatabase;
     messaging: () => FirebaseMessaging;
-    performance: () => any; // SEMVER: once >= 6 import performance.Performance
+    performance: () => FirebasePerformance;
     storage: (storageBucket?: string) => FirebaseStorage;
     delete: () => Promise<void>;
     firestore: () => FirebaseFirestore;
     functions: (region?: string) => FirebaseFunctions;
+    remoteConfig: () => FirebaseRemoteConfig;
     registerVersion?: (library: string, version: string) => void;
 }
 
 export const VERSION = new Version('ANGULARFIRE2_VERSION');
 
-export function _firebaseAppFactory(options: FirebaseOptions, nameOrConfig?: string|FirebaseAppConfig|null) {
+export function _firebaseAppFactory(options: FirebaseOptions, zone: NgZone, nameOrConfig?: string|FirebaseAppConfig|null) {
     const name = typeof nameOrConfig === 'string' && nameOrConfig || '[DEFAULT]';
     const config = typeof nameOrConfig === 'object' && nameOrConfig || {};
     config.name = config.name || name;
@@ -43,7 +55,7 @@ export function _firebaseAppFactory(options: FirebaseOptions, nameOrConfig?: str
     const existingApp = firebase.apps.filter(app => app && app.name === config.name)[0] as any;
     // We support FirebaseConfig, initializeApp's public type only accepts string; need to cast as any
     // Could be solved with https://github.com/firebase/firebase-js-sdk/pull/1206
-    const app = (existingApp || firebase.initializeApp(options, config as any)) as FirebaseApp;
+    const app = (existingApp || zone.runOutsideAngular(() => firebase.initializeApp(options, config as any))) as FirebaseApp;
     if (app.registerVersion) {
         app.registerVersion('angularfire', VERSION.full);
         app.registerVersion('angular', NG_VERSION.full);
@@ -55,8 +67,9 @@ const FirebaseAppProvider = {
     provide: FirebaseApp,
     useFactory: _firebaseAppFactory,
     deps: [
-        FirebaseOptionsToken,
-        [new Optional(), FirebaseNameOrConfigToken]
+        FIREBASE_OPTIONS,
+        NgZone,
+        [new Optional(), FIREBASE_APP_NAME]
     ]
 };
  
@@ -68,8 +81,8 @@ export class AngularFireModule {
         return {
             ngModule: AngularFireModule,
             providers: [
-                { provide: FirebaseOptionsToken, useValue: options },
-                { provide: FirebaseNameOrConfigToken, useValue: nameOrConfig }
+                { provide: FIREBASE_OPTIONS, useValue: options },
+                { provide: FIREBASE_APP_NAME, useValue: nameOrConfig }
             ]
         }
     }