+++ src/functions/functions.ts
@@ -1,13 +1,17 @@
 import { Injectable, Inject, Optional, NgZone, PLATFORM_ID, InjectionToken } from '@angular/core';
 import { Observable, from } from 'rxjs';
-import { map } from 'rxjs/operators';
-import { FirebaseOptions, FirebaseAppConfig } from '@angular/fire';
-import { FirebaseFunctions, FirebaseOptionsToken, FirebaseNameOrConfigToken, _firebaseAppFactory, FirebaseZoneScheduler } from '@angular/fire';
+import { map, observeOn } from 'rxjs/operators';
+import { FirebaseOptions, FirebaseAppConfig, FIREBASE_APP_NAME } from '@angular/fire';
+import { FirebaseFunctions, FIREBASE_OPTIONS, _firebaseAppFactory, ɵAngularFireSchedulers } from '@angular/fire';
 
-// SEMVER: @ v6 remove FunctionsRegionToken in favor of FUNCTIONS_REGION
+// SEMVER: @ v6 remove FunctionsRegionToken and FUNCTIONS_REGION in favor of REGION
 export const FunctionsRegionToken = new InjectionToken<string>('angularfire2.functions.region');
-export const FUNCTIONS_ORIGIN = new InjectionToken<string>('angularfire2.functions.origin');
 export const FUNCTIONS_REGION = FunctionsRegionToken;
+// SEMVER: @ v6 remove FUNCTIONS_ORIGIN in favor of ORIGIN
+export const FUNCTIONS_ORIGIN = new InjectionToken<string>('angularfire2.functions.origin');
+
+export const ORIGIN = FUNCTIONS_ORIGIN;
+export const REGION = FunctionsRegionToken;
 
 @Injectable()
 export class AngularFireFunctions {
@@ -17,20 +21,20 @@ export class AngularFireFunctions {
    */
   public readonly functions: FirebaseFunctions;
 
-  public readonly scheduler: FirebaseZoneScheduler;
+  public readonly schedulers: ɵAngularFireSchedulers;
 
   constructor(
-    @Inject(FirebaseOptionsToken) options:FirebaseOptions,
-    @Optional() @Inject(FirebaseNameOrConfigToken) nameOrConfig:string|FirebaseAppConfig|null|undefined,
+    @Inject(FIREBASE_OPTIONS) options:FirebaseOptions,
+    @Optional() @Inject(FIREBASE_APP_NAME) nameOrConfig:string|FirebaseAppConfig|null|undefined,
     @Inject(PLATFORM_ID) platformId: Object,
     zone: NgZone,
-    @Optional() @Inject(FUNCTIONS_REGION) region:string|null,
-    @Optional() @Inject(FUNCTIONS_ORIGIN) origin:string|null
+    @Optional() @Inject(REGION) region:string|null,
+    @Optional() @Inject(ORIGIN) origin:string|null
   ) {
-    this.scheduler = new FirebaseZoneScheduler(zone, platformId);
-    
+    this.schedulers = new ɵAngularFireSchedulers(zone);
+
     this.functions = zone.runOutsideAngular(() => {
-      const app = _firebaseAppFactory(options, nameOrConfig);
+      const app = _firebaseAppFactory(options, zone, nameOrConfig);
       return app.functions(region || undefined);
     });
 
@@ -44,11 +48,10 @@ export class AngularFireFunctions {
     const callable = this.functions.httpsCallable(name);
     return (data: T) => {
       const callable$ = from(callable(data));
-      return this.scheduler.runOutsideAngular(
-        callable$.pipe(
-          map(r => r.data as R)
-        )
-      )
+      return callable$.pipe(
+        observeOn(this.schedulers.outsideAngular),
+        map(r => r.data as R)
+      );
     }
   }