+++ src/storage/storage.ts
@@ -2,11 +2,13 @@ import { Injectable, Inject, Optional, InjectionToken, NgZone, PLATFORM_ID } fro
 import { createStorageRef, AngularFireStorageReference } from './ref';
 import { createUploadTask, AngularFireUploadTask } from './task';
 import { Observable } from 'rxjs';
-import { FirebaseStorage, FirebaseOptions, FirebaseAppConfig, FirebaseOptionsToken, FirebaseNameOrConfigToken, FirebaseZoneScheduler, _firebaseAppFactory } from '@angular/fire';
+import { FirebaseStorage, FirebaseOptions, FirebaseAppConfig, _firebaseAppFactory, FIREBASE_OPTIONS, FIREBASE_APP_NAME, ɵkeepUnstableUntilFirstFactory, ɵAngularFireSchedulers } from '@angular/fire';
 
 import { UploadMetadata } from './interfaces';
 
+// SEMVER drop StorageBucket in favor of BUCKET
 export const StorageBucket = new InjectionToken<string>('angularfire2.storageBucket');
+export const BUCKET = StorageBucket;
 
 /**
  * AngularFireStorage Service
@@ -18,29 +20,33 @@ export const StorageBucket = new InjectionToken<string>('angularfire2.storageBuc
 @Injectable()
 export class AngularFireStorage {
   public readonly storage: FirebaseStorage;
-  public readonly scheduler: FirebaseZoneScheduler;
+
+  public readonly keepUnstableUntilFirst: <T>(obs: Observable<T>) => Observable<T>;
+  public readonly schedulers: ɵAngularFireSchedulers;
 
   constructor(
-    @Inject(FirebaseOptionsToken) options:FirebaseOptions,
-    @Optional() @Inject(FirebaseNameOrConfigToken) nameOrConfig:string|FirebaseAppConfig|null|undefined,
-    @Optional() @Inject(StorageBucket) storageBucket:string|null,
+    @Inject(FIREBASE_OPTIONS) options:FirebaseOptions,
+    @Optional() @Inject(FIREBASE_APP_NAME) nameOrConfig:string|FirebaseAppConfig|null|undefined,
+    @Optional() @Inject(BUCKET) storageBucket:string|null,
     @Inject(PLATFORM_ID) platformId: Object,
     zone: NgZone
   ) {
-    this.scheduler = new FirebaseZoneScheduler(zone, platformId);
+    this.schedulers = new ɵAngularFireSchedulers(zone);
+    this.keepUnstableUntilFirst = ɵkeepUnstableUntilFirstFactory(this.schedulers, platformId);
+
     this.storage = zone.runOutsideAngular(() => {
-      const app = _firebaseAppFactory(options, nameOrConfig);
+      const app = _firebaseAppFactory(options, zone, nameOrConfig);
       return app.storage(storageBucket || undefined);
     });
   }
 
   ref(path: string) {
-    return createStorageRef(this.storage.ref(path), this.scheduler);
+    return createStorageRef(this.storage.ref(path), this.schedulers, this.keepUnstableUntilFirst);
   }
 
   upload(path: string, data: any, metadata?: UploadMetadata) {
     const storageRef = this.storage.ref(path);
-    const ref = createStorageRef(storageRef, this.scheduler);
+    const ref = createStorageRef(storageRef, this.schedulers, this.keepUnstableUntilFirst);
     return ref.put(data, metadata);
   }