+++ tools/build.js
@@ -5,6 +5,7 @@ const { switchMap, switchMapTo, tap } = require('rxjs/operators');
 const { copy, readFileSync, writeFile, writeFileSync, statSync } = require('fs-extra');
 const { prettySize } = require('pretty-size');
 const gzipSize = require('gzip-size');
+const path = require('path');
 const resolve = require('rollup-plugin-node-resolve');
 const pkg = require(`${process.cwd()}/package.json`);
 
@@ -21,6 +22,7 @@ const GLOBALS = {
   '@angular-devkit/core/node': 'ng-devkit.core.node',
   '@angular-devkit/architect': 'ng-devkit.architect',
   'firebase': 'firebase',
+  'firebase/analytics': 'firebase',
   'firebase/app': 'firebase',
   'firebase/auth': 'firebase',
   'firebase/database': 'firebase',
@@ -28,9 +30,12 @@ const GLOBALS = {
   'firebase/firestore': 'firebase',
   'firebase/functions': 'firebase',
   'firebase/performance': 'firebase',
+  'firebase/remote-config': 'firebase',
+  '@firebase/remote-config': 'firebase',
   'firebase/storage': 'firebase',
   '@angular/fire': 'angularfire2',
   '@angular/fire/auth': 'angularfire2.auth',
+  '@angular/fire/analytics': 'angularfire2.analytics',
   '@angular/fire/auth-guard': 'angularfire2.auth_guard',
   '@angular/fire/database': 'angularfire2.database',
   '@angular/fire/database-deprecated': 'angularfire2.database_deprecated',
@@ -39,6 +44,7 @@ const GLOBALS = {
   '@angular/fire/storage': 'angularfire2.storage',
   '@angular/fire/messaging': 'angularfire2.messaging',
   '@angular/fire/performance': 'angularfire2.performance',
+  '@angular/fire/remote-config': 'angularfire2.remote_config',
   'fs': 'node.fs',
   'path': 'node.path',
   'inquirer': 'inquirer'
@@ -65,6 +71,7 @@ const VERSIONS = {
 
 const MODULE_NAMES = {
   core: 'angularfire2',
+  analytics: 'angularfire2.analytics',
   auth: 'angularfire2.auth',
   "auth-guard": 'angularfire2.auth_guard',
   database: 'angularfire2.database',
@@ -74,11 +81,13 @@ const MODULE_NAMES = {
   schematics: 'angularfire2.schematics',
   storage: 'angularfire2.storage',
   messaging: 'angularfire2.messaging',
-  performance: 'angularfire2.performance'
+  performance: 'angularfire2.performance',
+  "remote-config": 'angularfire2.remote_config'
 };
 
 const ENTRIES = {
   core: `${process.cwd()}/dist/packages-dist/index.js`,
+  analytics: `${process.cwd()}/dist/packages-dist/analytics/index.js`,
   auth: `${process.cwd()}/dist/packages-dist/auth/index.js`,
   "auth-guard": `${process.cwd()}/dist/packages-dist/auth-guard/index.js`,
   database: `${process.cwd()}/dist/packages-dist/database/index.js`,
@@ -88,11 +97,13 @@ const ENTRIES = {
   schematics: `${process.cwd()}/dist/packages-dist/schematics/index.js`,
   storage: `${process.cwd()}/dist/packages-dist/storage/index.js`,
   messaging: `${process.cwd()}/dist/packages-dist/messaging/index.js`,
-  performance: `${process.cwd()}/dist/packages-dist/performance/index.js`
+  performance: `${process.cwd()}/dist/packages-dist/performance/index.js`,
+  "remote-config": `${process.cwd()}/dist/packages-dist/remote-config/index.js`
 };
 
 const SRC_PKG_PATHS = {
   core: `${process.cwd()}/src/core/package.json`,
+  analytics: `${process.cwd()}/src/analytics/package.json`,
   auth: `${process.cwd()}/src/auth/package.json`,
   "auth-guard": `${process.cwd()}/src/auth-guard/package.json`,
   database: `${process.cwd()}/src/database/package.json`,
@@ -103,12 +114,14 @@ const SRC_PKG_PATHS = {
   storage: `${process.cwd()}/src/storage/package.json`,
   messaging: `${process.cwd()}/src/messaging/package.json`,
   performance: `${process.cwd()}/src/performance/package.json`,
+  "remote-config": `${process.cwd()}/src/remote-config/package.json`,
   schematics: `${process.cwd()}/dist/packages-dist/schematics/versions.js`,
   "schematics-es2015": `${process.cwd()}/dist/packages-dist/schematics/es2015/versions.js`
 };
 
 const DEST_PKG_PATHS = {
   core: `${process.cwd()}/dist/packages-dist/package.json`,
+  analytics: `${process.cwd()}/dist/packages-dist/analytics/package.json`,
   auth: `${process.cwd()}/dist/packages-dist/auth/package.json`,
   "auth-guard": `${process.cwd()}/dist/packages-dist/auth-guard/package.json`,
   database: `${process.cwd()}/dist/packages-dist/database/package.json`,
@@ -119,13 +132,16 @@ const DEST_PKG_PATHS = {
   storage: `${process.cwd()}/dist/packages-dist/storage/package.json`,
   messaging: `${process.cwd()}/dist/packages-dist/messaging/package.json`,
   performance: `${process.cwd()}/dist/packages-dist/performance/package.json`,
+  "remote-config": `${process.cwd()}/dist/packages-dist/remote-config/package.json`,
   schematics: `${process.cwd()}/dist/packages-dist/schematics/versions.js`,
   "schematics-es2015": `${process.cwd()}/dist/packages-dist/schematics/es2015/versions.js`
 };
 
 // Constants for running typescript commands
-const TSC = 'node_modules/.bin/tsc';
-const NGC = 'node_modules/.bin/ngc';
+const binSuffix = process.platform === "win32" ? '.cmd' : '';
+const TSC = path.normalize('node_modules/.bin/tsc' + binSuffix);
+const NGC = path.normalize('node_modules/.bin/ngc' +  binSuffix);
+
 const TSC_ARGS = (name, config = 'build') => [`-p`, `${process.cwd()}/src/${name}/tsconfig-${config}.json`];
 const TSC_TEST_ARGS = [`-p`, `${process.cwd()}/src/tsconfig-test.json`];
 
@@ -283,8 +299,12 @@ function copySchematicFiles() {
 }
 
 function replaceDynamicImportsForUMD() {
-  writeFileSync('./dist/packages-dist/bundles/performance.umd.js', readFileSync('./dist/packages-dist/bundles/performance.umd.js', 'utf8').replace("rxjs.from(import('firebase/performance'))", "rxjs.empty()"));
+  writeFileSync('./dist/packages-dist/bundles/performance.umd.js', readFileSync('./dist/packages-dist/bundles/performance.umd.js', 'utf8').replace("return import('firebase/performance');", "return rxjs.empty();"));
   writeFileSync('./dist/packages-dist/bundles/messaging.umd.js', readFileSync('./dist/packages-dist/bundles/messaging.umd.js', 'utf8').replace("rxjs.from(import('firebase/messaging'))", "rxjs.empty()"));
+  writeFileSync('./dist/packages-dist/bundles/remote-config.umd.js', readFileSync('./dist/packages-dist/bundles/remote-config.umd.js', 'utf8').replace("return import('firebase/remote-config');", "return rxjs.empty();"));
+  writeFileSync('./dist/packages-dist/bundles/analytics.umd.js', readFileSync('./dist/packages-dist/bundles/analytics.umd.js', 'utf8').replace("return import('firebase/analytics');", "return rxjs.empty();"));
+  // TODO move into own step
+  writeFileSync('./dist/packages-dist/remote-config/remote-config.d.ts', readFileSync('./dist/packages-dist/remote-config/remote-config.d.ts', 'utf8').replace("implements remoteConfig.Value", "implements Partial<remoteConfig.Value>"));
 }
 
 function writeCoreVersion() {
@@ -308,6 +328,7 @@ function measure(module) {
 function getVersions() {
   const paths = [
     getDestPackageFile('core'),
+    getDestPackageFile('analytics'),
     getDestPackageFile('auth'),
     getDestPackageFile('auth-guard'),
     getDestPackageFile('database'),
@@ -317,6 +338,7 @@ function getVersions() {
     getDestPackageFile('storage'),
     getDestPackageFile('messaging'),
     getDestPackageFile('performance'),
+    getDestPackageFile('remote-config'),
     getDestPackageFile('database-deprecated')
   ];
   return paths
@@ -350,6 +372,7 @@ function buildModule(name, globals) {
  */
 function buildModules(globals) {
   const core$ = buildModule('core', globals);
+  const analytics$ = buildModule('analytics', globals);
   const auth$ = buildModule('auth', globals);
   const authGuard$ = buildModule('auth-guard', globals);
   const db$ = buildModule('database', globals);
@@ -359,9 +382,11 @@ function buildModules(globals) {
   const storage$ = buildModule('storage', globals);
   const messaging$ = buildModule('messaging', globals);
   const performance$ = buildModule('performance', globals);
+  const remoteConfig$ = buildModule('remote-config', globals);
   const dbdep$ = buildModule('database-deprecated', globals);
   return forkJoin(core$, from(copyRootTest())).pipe(
     switchMapTo(schematics$),
+    switchMapTo(analytics$),
     switchMapTo(auth$),
     switchMapTo(authGuard$),
     switchMapTo(db$),
@@ -370,6 +395,7 @@ function buildModules(globals) {
     switchMapTo(storage$),
     switchMapTo(messaging$),
     switchMapTo(performance$),
+    switchMapTo(remoteConfig$),
     switchMapTo(dbdep$)
   );
 }
@@ -390,6 +416,7 @@ function buildLibrary(globals) {
       replaceDynamicImportsForUMD();
       writeCoreVersion();
       const coreStats = measure('core');
+      const analyticsStats = measure('analytics');
       const authStats = measure('auth');
       const authGuardStats = measure('auth-guard');
       const dbStats = measure('database');
@@ -398,9 +425,11 @@ function buildLibrary(globals) {
       const storageStats = measure('storage');
       const messagingStats = measure('messaging');
       const performanceStats = measure('performance');
+      const remoteConfigStats = measure('remote-config');
       const dbdepStats = measure('database-deprecated');
       console.log(`
 core.umd.js - ${coreStats.size}, ${coreStats.gzip}
+analytics.umd.js - ${analyticsStats.size}, ${analyticsStats.gzip}
 auth.umd.js - ${authStats.size}, ${authStats.gzip}
 auth-guard.umd.js - ${authGuardStats.size}, ${authGuardStats.gzip}
 database.umd.js - ${dbStats.size}, ${dbStats.gzip}
@@ -409,6 +438,7 @@ functions.umd.js - ${functionsStats.size}, ${functionsStats.gzip}
 storage.umd.js - ${storageStats.size}, ${storageStats.gzip}
 messaging.umd.js - ${messagingStats.size}, ${messagingStats.gzip}
 performance.umd.js - ${performanceStats.size}, ${performanceStats.gzip}
+remote-config.umd.js - ${remoteConfigStats.size}, ${remoteConfigStats.gzip}
 database-deprecated.umd.js - ${dbdepStats.size}, ${dbdepStats.gzip}
       `);
       verifyVersions();